Refactor to ControlHandler class
